requirements.txt
bluepy
pycrypto
curses-menu
crc16
pyserial
